SUDOKU-1 Change digit type from short to byte.

+++ SudokuGenerator/Data/Enum/Digit.cs
@@ -1,6 +1,6 @@
 ﻿namespace SudokuGenerator.Data.Enum
 {
-	public enum Digit : short
+	public enum Digit : byte
 	{
 		None = 0,
 		One = 1,
